Korean Festivals Data

festival_data returns the raw data frame containing detailed information about festivals in Korea, including names, locations, dates, and additional details in both Korean and English. English translations are done through =googletranslate() on Google Sheets, and as such may not be accurate.

Usage

festival_data

Format

A data frame with 21 variables:

festival_kr

The name of the festival in Korean.

festival_en

The English translation of the festival's name.

place

The place where the festival is held, in Korean.

place_en

The English translation of the festival's location.

begin_date

The start date of the festival.

end_date

The end date of the festival.

contents

Description of the festival in Korean.

contents_en

The English translation of the festival's description.

organizer

The organizer of the festival in Korean.

organizer_en

The English translation of the organizer's name.

address

The address of the festival location in Korean.

address_en

The English translation of the address.

address_old_style

The old-style address in Korean.

address_old_en

The English translation of the old-style address.

latitude

Latitude coordinate of the festival location.

longitude

Longitude coordinate of the festival location.

updated

The date when the data was last updated.

data_organization_code

Code of the organization providing the data.

data_organization

The organization providing the data, in Korean.

data_organization_en

The English translation of the data organization's name.

other columns

Descriptions for any additional columns not listed here.

Source

https://www.data.go.kr/data/15013104/standard.do

Value

Returns the raw data frame containing information on Korean festivals